% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_nba.R
\name{ContentToDataFrame}
\alias{ContentToDataFrame}
\title{NBA Content to DataFrame}
\usage{
ContentToDataFrame(content, ix)
}
\value{
data frame with info from stats.nba.com
}
\description{
NBA Content to DataFrame
}
