% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreeAgentProp.R
\name{freeAgentProp}
\alias{freeAgentProp}
\title{provide free agent signing proposal with given team}
\usage{
freeAgentProp(team, cap = 1.21e+08, accuracy = 50)
}
\arguments{
\item{team}{(list of players)}

\item{cap}{(salary cap, default $121000000, which is the luxury tax line)}

\item{accuracy}{(default 50, run time is directly proportion to accuracy)}
}
\value{
no return
}
\description{
provide free agent signing proposal with given team
}
