% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GamePred.R
\name{gamePred}
\alias{gamePred}
\title{predict game result with team and oppteam}
\usage{
gamePred(team, opp_team)
}
\arguments{
\item{team:}{c(player names)}

\item{opp_team:}{c(player names in opposite team)}
}
\value{
game result prediction: W / L
}
\description{
predict game result with team and oppteam
}
