% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_scrape.R
\name{GetBoxScore}
\alias{GetBoxScore}
\title{Box Score}
\usage{
GetBoxScore(...)
}
\value{
data frame with game's box score
}
\description{
Box Score
}
\examples{
# GetBoxScore(GameID = '0021600794')
}
\keyword{boxscore}
