% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_scrape.R
\name{GetGameLogs}
\alias{GetGameLogs}
\title{Game Logs}
\usage{
GetGameLogs(...)
}
\value{
data frame with game logs for all players or teams
}
\description{
Game Logs
}
\examples{
# GetGameLogs(SeasonType = 'Playoffs')
}
\keyword{gamelog}
