% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreboard_scrape.R
\name{GetGames}
\alias{GetGames}
\title{Game Schedule}
\usage{
GetGames(...)
}
\value{
data frame with list of games on a certain day
}
\description{
Game Schedule
}
\examples{
# GetGames(Date = '2017-02-02')
}
\keyword{games}
