% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{GetLineupStats}
\alias{GetLineupStats}
\title{Lineup Stats}
\usage{
GetLineupStats(...)
}
\value{
data frame with stats for all lineups
}
\description{
Lineup Stats
}
\examples{
GetLineupStats(SeasonType = 'Playoffs')
}
\keyword{lineup}
