% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_scrape.R
\name{GetPlayByPlay}
\alias{GetPlayByPlay}
\title{Play by play}
\usage{
GetPlayByPlay(...)
}
\value{
data frame with list of plays in a game
}
\description{
Play by play
}
\examples{
# GetPlayByPlay(GameID = '0021600794')
}
\keyword{plays}
