% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_scrape.R
\name{GetPlayerBioStats}
\alias{GetPlayerBioStats}
\title{Player Bio Stats}
\usage{
GetPlayerBioStats(...)
}
\value{
data frame with bio stats for all players
}
\description{
Player Bio Stats
}
\examples{
# GetPlayerBioStats(SeasonType = 'Playoffs')
}
\keyword{bio}
\keyword{player}
