% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{GetPlayerClutchStats}
\alias{GetPlayerClutchStats}
\title{Player Clutch Stats}
\usage{
GetPlayerClutchStats(...)
}
\value{
data frame with stats for all players in the clutch
}
\description{
Player Clutch Stats
}
\examples{
GetPlayerClutchStats(SeasonType = 'Playoffs')
}
\keyword{clutch}
\keyword{player}
