% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_scrape.R
\name{GetPlayerDashboardGameSplits}
\alias{GetPlayerDashboardGameSplits}
\title{Player dashboard game splits}
\usage{
GetPlayerDashboardGameSplits(split = "Areas", ...)
}
\value{
data frame with players dashboard game splits
}
\description{
Player dashboard game splits
}
\examples{
# GetPlayerDashboardGameSplits(PlayerID = '2544')
}
\keyword{dashboard}
\keyword{player}
\keyword{splits}
