% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_scrape.R
\name{GetPlayerDashboardGeneralSplits}
\alias{GetPlayerDashboardGeneralSplits}
\title{Player dashboard general splits}
\usage{
GetPlayerDashboardGeneralSplits(split = "Areas", ...)
}
\value{
data frame with players dashboard general splits
}
\description{
Player dashboard general splits
}
\examples{
# GetPlayerDashboardGeneralSplits(PlayerID = '2544')
}
\keyword{dashboard}
\keyword{player}
\keyword{splits}
