% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_scrape.R
\name{GetPlayerDashboardSplits}
\alias{GetPlayerDashboardSplits}
\title{Player dashboard splits}
\usage{
GetPlayerDashboardSplits(split = "Areas", ...)
}
\value{
data frame with players general dashboard splits
}
\description{
Player dashboard splits
}
\examples{
# GetPlayerDashboardSplits(PlayerID = '2544')
}
\keyword{dashboard}
\keyword{player}
\keyword{splits}
