% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_scrape.R
\name{GetPlayerDefenseStats}
\alias{GetPlayerDefenseStats}
\title{Defense Stats}
\usage{
GetPlayerDefenseStats(...)
}
\value{
data frame with defensive stats for all players
}
\description{
Defense Stats
}
\examples{
# GetPlayerDefenseStats(SeasonType = 'Playoffs')
}
\keyword{player}
