% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_scrape.R
\name{GetPlayerGameLogs}
\alias{GetPlayerGameLogs}
\title{Player game logs}
\usage{
GetPlayerGameLogs(...)
}
\value{
data frame with all games for a player
}
\description{
Player game logs
}
\examples{
# GetPlayerGameLogs()
}
\keyword{game}
\keyword{log}
\keyword{player}
