% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_scrape.R
\name{GetPlayerShootingSplits}
\alias{GetPlayerShootingSplits}
\title{Player shooting splits}
\usage{
GetPlayerShootingSplits(split = "Areas", ...)
}
\value{
data frame with players shooting splits
}
\description{
Player shooting splits
}
\examples{
# GetPlayerShootingSplits(PlayerID = '2544')
}
\keyword{player}
\keyword{shooting}
\keyword{splits}
