% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_scrape.R
\name{GetPlayerShootingStats}
\alias{GetPlayerShootingStats}
\title{Player shooting}
\usage{
GetPlayerShootingStats(...)
}
\value{
data frame with shooting stats for all players
}
\description{
Player shooting
}
\examples{
# GetPlayerShootingStats(SeasonType = 'Playoffs')
}
\keyword{player}
\keyword{shooting}
