% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_scrape.R
\name{GetPlayerShotLocationStats}
\alias{GetPlayerShotLocationStats}
\title{Player Shot Location Stats}
\usage{
GetPlayerShotLocationStats(...)
}
\value{
data frame with shot location stats for all players
}
\description{
Player Shot Location Stats
}
\examples{
# GetPlayerShotLocationStats(SeasonType = 'Playoffs')
}
\keyword{location}
\keyword{player}
\keyword{shot}
