% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_scrape.R
\name{GetPlayerYearByYearStats}
\alias{GetPlayerYearByYearStats}
\title{Player year over year stats}
\usage{
GetPlayerYearByYearStats(split = "Advanced", source = "NBA", ...)
}
\value{
data frame with players yearly stats
}
\description{
Player year over year stats
}
\examples{
# GetPlayerYearByYearStats(PlayerID = '2544')
}
\keyword{game}
\keyword{log}
\keyword{player}
