% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_scrape.R
\name{GetPlayTypeStats}
\alias{GetPlayTypeStats}
\title{Play Type Stats}
\usage{
GetPlayTypeStats(type = "player", ...)
}
\value{
data frame with play type stats for all players or all teams
}
\description{
Play Type Stats
}
\examples{
# GetPlayTypeStats(category = 'Isolation')
}
\keyword{play}
\keyword{player}
\keyword{team}
\keyword{type}
