% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{GetTeamGeneralSplits}
\alias{GetTeamGeneralSplits}
\title{Team Dashboard by General Splits}
\usage{
GetTeamGeneralSplits(split = "location", ...)
}
\value{
data frame with team information by a variety of splits
}
\description{
Team Dashboard by General Splits
}
\examples{
GetTeamGeneralSplits(split = 'location', TeamID = '1610612756')
}
\keyword{dashboard}
\keyword{splits}
\keyword{team}
