% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_scrape.R
\name{GetTeamHistory}
\alias{GetTeamHistory}
\title{Team History}
\usage{
GetTeamHistory(source = "BRef", ...)
}
\value{
data frame with historical records for franchise
}
\description{
Team History
}
\examples{
# GetTeamHistory(TeamID = 'HOU')
}
\keyword{history}
\keyword{team}
