% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{GetTeamLogo}
\alias{GetTeamLogo}
\title{Get Team Logo}
\usage{
GetTeamLogo(team.id)
}
\arguments{
\item{team.id}{Team ID from ESPN (e.g. 'hou')}
}
\value{
team's logo
}
\description{
Get Team Logo
}
\examples{
GetTeamLogo('hou')
}
\keyword{logo}
\keyword{picture}
\keyword{team}
