% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_scrape.R
\name{GetTeamPlayerDashboard}
\alias{GetTeamPlayerDashboard}
\title{Team player stats}
\usage{
GetTeamPlayerDashboard(...)
}
\value{
data frame with stats for players on a team
}
\description{
Team player stats
}
\examples{
# GetTeamPlayerDashboard(SeasonType = 'Playoffs', 'TeamID' = '1610612745')
}
\keyword{player}
\keyword{team}
