% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{GetTeamPlayerOnOffDetails}
\alias{GetTeamPlayerOnOffDetails}
\title{Team On Off Details}
\usage{
GetTeamPlayerOnOffDetails(...)
}
\value{
data frame with on-off stats for players on a team
}
\description{
Team On Off Details
}
\examples{
GetTeamPlayerOnOffDetails(SeasonType = 'Playoffs')
}
\keyword{on-off}
\keyword{team}
