% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_scrape.R
\name{GetTeamPossessions}
\alias{GetTeamPossessions}
\title{Possession Finder}
\usage{
GetTeamPossessions(ix = "team_results", ...)
}
\value{
data frame with possessions
}
\description{
Possession Finder
}
\examples{
# GetTeamPossessions(TeamID = 'HOU')
}
\keyword{possessions}
