% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_scrape.R
\name{GetTeamRoster}
\alias{GetTeamRoster}
\title{Team player stats}
\usage{
GetTeamRoster(...)
}
\value{
data frame with stats for players on a team
}
\description{
Team player stats
}
\examples{
# GetTeamRoster('TeamID' = '1610612745')
}
\keyword{player}
\keyword{team}
