% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_scrape.R
\name{GetTeamShootingStats}
\alias{GetTeamShootingStats}
\title{Team shooting}
\usage{
GetTeamShootingStats(...)
}
\value{
data frame with shooting stats for all teams
}
\description{
Team shooting
}
\examples{
# GetTeamShootingStats(SeasonType = 'Playoffs')
}
\keyword{shooting}
\keyword{team}
