% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_scrape.R
\name{GetTeamShotDashboard}
\alias{GetTeamShotDashboard}
\title{Team Shot Dashboard}
\usage{
GetTeamShotDashboard(split = "shot clock", ...)
}
\value{
data frame with team information by a variety of splits
}
\description{
Team Shot Dashboard
}
\examples{
# GetTeamShotDashboard(split = 'shot clock', TeamID = '1610612756')
}
\keyword{dashboard}
\keyword{shot}
\keyword{team}
