% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_scrape.R
\name{GetTeamShotLocationStats}
\alias{GetTeamShotLocationStats}
\title{Team shot location}
\usage{
GetTeamShotLocationStats(...)
}
\value{
data frame with shot location stats for all teams
}
\description{
Team shot location
}
\examples{
# GetTeamShotLocationStats(SeasonType = 'Playoffs')
}
\keyword{location}
\keyword{shot}
\keyword{team}
