% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{GetTeamStats}
\alias{GetTeamStats}
\title{Team Stats}
\usage{
GetTeamStats(source = "NBA", ...)
}
\value{
data frame with stats for all teams
}
\description{
Team Stats
}
\examples{
GetTeamStats(SeasonType = 'Playoffs')
}
\keyword{team}
