% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_scrape.R
\name{GetTrackingStats}
\alias{GetTrackingStats}
\title{Tracking Stats}
\usage{
GetTrackingStats(...)
}
\value{
data frame with tracking stats for all players or all teams
}
\description{
Tracking Stats
}
\examples{
# GetTrackingStats(PlayerOrTeam = 'Team')
}
\keyword{player}
\keyword{team}
\keyword{tracking}
