% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_scrape.R
\name{GetWOWYStats}
\alias{GetWOWYStats}
\title{WOWY Stats}
\usage{
GetWOWYStats(...)
}
\value{
data frame with wowy stats (e.g. Steph on KD off)
}
\description{
WOWY Stats
}
\examples{
# GetWOWYStats(TeamID = 'HOU')
}
\keyword{wowy}
