% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{playersToTeamData}
\alias{playersToTeamData}
\title{calculate team stats with players stats}
\usage{
playersToTeamData(team)
}
\arguments{
\item{team:}{c(player names)}
}
\value{
team stats
}
\description{
calculate team stats with players stats
}
