% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SalaryPred.R
\name{salaryPred}
\alias{salaryPred}
\title{predict salary according to player's performance}
\usage{
salaryPred(player)
}
\arguments{
\item{player:}{player name}
}
\value{
salary prediction(in dollar)
}
\description{
predict salary according to player's performance
}
