% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamQualityPred.R
\name{teamQualityPred}
\alias{teamQualityPred}
\title{predict team quality according to player's performance}
\usage{
teamQualityPred(team, echo = 1000)
}
\arguments{
\item{team:}{c(player names)}

\item{echo:}{number of simulating games}
}
\value{
team quality(60 ~ 100)
}
\description{
predict team quality according to player's performance
}
