% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeProp.R
\name{tradeProp}
\alias{tradeProp}
\title{provide proposal of a certain trade}
\usage{
tradeProp(team_bt, team_at, accuracy = 1000)
}
\arguments{
\item{team_bt}{(team before trade, list of players)}

\item{team_at}{(team after trade, list of players)}

\item{accuracy}{(default 1000)}
}
\value{
no return
}
\description{
provide proposal of a certain trade
}
