% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{GetLineups}
\alias{GetLineups}
\title{Lineups}
\usage{
GetLineups(GameID)
}
\value{
Data frame with 5 man lineups with play-by-play from NBA.com
}
\description{
Lineups
}
\examples{
# GetLineups(GameID = '0021300359')
}
\keyword{lineups}
\keyword{playbyplay}
