% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_scrape.R
\name{GetPlayerStats}
\alias{GetPlayerStats}
\title{Player Stats}
\usage{
GetPlayerStats(source = "NBA", ...)
}
\value{
data frame with stats for all players
}
\description{
Player Stats
}
\examples{
# GetPlayerStats(SeasonType = 'Playoffs')
}
\keyword{player}
