% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{GetTeamGameLog}
\alias{GetTeamGameLog}
\title{Team game log}
\usage{
GetTeamGameLog(...)
}
\value{
data frame with game box scores for a team
}
\description{
Team game log
}
\examples{
GetTeamGameLog(SeasonType = 'Playoffs', 'TeamID' = '1610612745')
}
\keyword{player}
